-- cors should be fixed

+++ src/main/java/com/bonfire/az/bonfireaz/conf/UserDetailsServiceJPA.java
@@ -1,7 +1,7 @@
 package com.bonfire.az.bonfireaz.conf;
 
-import com.bonfire.az.bonfireaz.entity.db.XUser;
-import com.bonfire.az.bonfireaz.entity.sec.XUserDetails;
+import com.bonfire.az.bonfireaz.model.entity.XUser;
+import com.bonfire.az.bonfireaz.model.sec.XUserDetails;
 import com.bonfire.az.bonfireaz.repo.UserRepo;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.context.annotation.Configuration;
@@ -39,7 +39,7 @@ public class UserDetailsServiceJPA implements UserDetailsService {
                 ));
     }
 
-    public UserDetails loadUserByUserId(int id) throws UsernameNotFoundException {
+    public UserDetails loadUserByUserId(long id) throws UsernameNotFoundException {
         log.info(String.format(">>>> loading user details for user(id): %s", id));
         return repo.findById(id)
                 .map(UserDetailsServiceJPA::mapper_to_xUserDetails)
